[ty] Move the type hierarchy request handlers to individual modules
